requirements.txt
google-api-python-client>=2.160.0
google-auth>=2.38.0
google-auth-oauthlib>=1.2.1
google-genai>=1.12.1
python-pptx>=0.6.23
